requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
livekit==0.17.3
livekit-api==0.7.1
livekit-agents==0.10.6
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-elevenlabs
python-dotenv
